debian setup script updated

+++ setup-debian/setup.sh
@@ -5,20 +5,26 @@ clear && sleep 3
 
 
 ## clean apt cache
+echo -e "\nINFO: cleaning existing package cache ...\n"
 sudo apt clean
 
 
 ## add non-free and contrib sources
+echo -e "\nINFO: setting up additional package sources ...\n"
 sudo apt-add-repository non-free && \
 sudo apt-add-repository contrib
 
 
 ## remove applications (optional, uncomment if needed)
+# echo -e "\nINFO: removing unused applications ...\n"
 # sudo apt remove -y --purge gnome-games gnome-contacts gnome-maps \
-# gnome-weather
+# gnome-weather && \
+# sudo apt autoremove
+
 
 
 ## configure icons and themes
+echo -e "\nINFO: installing additional themes and icons ...\n"
 sudo tar -xvf ./icons/macOSBigSur.tar.gz -C /usr/share/icons && \
 sudo tar -xvf ./icons/Pop_Cyan.tar.gz -C /usr/share/icons && \
 sudo tar -xvf ./icons/Pop_Cyan_Dark.tar.gz -C /usr/share/icons && \
@@ -29,14 +35,15 @@ sudo tar -xvf ./themes/Prof-Gnome-Light-DS-3.6.tar.xz -C /usr/share/themes
 
 
 ## install additional fonts
+echo -e "\nINFO: installing additional fonts ...\n"
 sudo cp -r ./fonts/* /usr/share/fonts/ && \
 sudo fc-cache -f -v
 
 
 ## define path valriables
-BKP_DIR="/home/koushik/.config-backup" # backup directory
-DL_PKGS_DIR="/home/koushik/Downloads/downloaded-packages" # directory to store downloaded packages
-MINICONDA_INSTALL_DIR="/home/koushik/miniconda" # miniconda installation directory
+BKP_DIR="/home/$USER/.config-backup" # backup directory
+DL_PKGS_DIR="/home/$USER/Downloads/downloaded-packages" # directory to store downloaded packages
+MINICONDA_INSTALL_DIR="/home/$USER/miniconda" # miniconda installation directory
 AUTOCPUFREQ_DIR="$DL_PKGS_DIR/auto-cpufreq/" # directory for cloning 'auto-cpufreq' from github
 
 
@@ -44,12 +51,14 @@ AUTOCPUFREQ_DIR="$DL_PKGS_DIR/auto-cpufreq/" # directory for cloning 'auto-cpufr
 if [ -d "$BKP_DIR" ]; then
     echo -e "\nINFO: $BKP_DIR already exists ...\n"
 else
+    echo -e "\nINFO: creating directory for backing up existing configuration files ...\n"
     mkdir $BKP_DIR && \
-    sudo chown -R koushik:koushik $BKP_DIR
+    sudo chown -R $USER:$USER $BKP_DIR
 fi
 
 
 ## install additional packages
+echo -e "\nINFO: configuring additional packages ...\n"
 sudo apt update && \
 sudo apt -y install git tlp powertop fonts-roboto \
 tlp-rdw gimp gnome-tweak-tool vlc curl wget synaptic \
@@ -62,8 +71,9 @@ dconf-editor timeshift p7zip-full p7zip-rar rar unrar
 if [ -d "$DL_PKGS_DIR" ]; then
     echo -e "\nINFO: $DL_PKGS_DIR already exists ...\n"
 else
+    echo -e "\nINFO: creating directory for downloading additional applications ...\n"
     mkdir $DL_PKGS_DIR && \
-    sudo chown -R koushik:koushik $DL_PKGS_DIR
+    sudo chown -R $USER:$USER $DL_PKGS_DIR
 fi
 
 
@@ -71,26 +81,34 @@ fi
 if [ -f "$DL_PKGS_DIR/miniconda.sh" ]; then
     echo -e "\nINFO: miniconda installation script available, avoiding download ...\n"
 else
+    echo -e "\nINFO: downloading miniconda installer script ...\n"
     wget https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh \
     -O "$DL_PKGS_DIR/miniconda.sh"
 fi
 
 if [ -d "$MINICONDA_INSTALL_DIR" ]; then
-    echo -e "\nINFO: $MINICONDA_INSTALL_DIR already exists, seems like 'miniconda' has already been configured ...\n" 
+    echo -e "\nINFO: $MINICONDA_INSTALL_DIR already exists, seems like 'miniconda' has already been installed ...\n" 
 else
+    echo -e "\nINFO: creating miniconda installation directory ...\n"
     bash "$DL_PKGS_DIR/miniconda.sh" -b -p $MINICONDA_INSTALL_DIR && \
-    sudo chown -R koushik:koushik $MINICONDA_INSTALL_DIR && \
-    cp /home/koushik/.bashrc "$BKP_DIR/.bashrc.bak" && \
-    echo -e "# miniconda configuration\nexport PATH="$MINICONDA_INSTALL_DIR/bin":$PATH" \
-    >> /home/koushik/.bashrc
+    sudo chown -R $USER:$USER $MINICONDA_INSTALL_DIR && \
+    cp /home/$USER/.bashrc "$BKP_DIR/.bashrc.bak" && \
+    echo -e "\n\n# miniconda configuration\nexport PATH="$MINICONDA_INSTALL_DIR/bin":$PATH" \
+    >> /home/$USER/.bashrc
 fi
 
 
 ## setup latest GNU R
-sudo apt-key add ./cran40/jranke.asc && \
-sudo cp ./cran40/gnu-r.list /etc/apt/sources.list.d/ && \
-sudo apt update && \
-apt install -y -t bullseye-cran40 r-base-dev
+if [ -f "/etc/apt/sources.list.d/gnu-r.list" ]; then
+    echo -e "\nINFO: seems like package sources have already been configured for GNU R ...\n"
+else
+    echo -e "\nINFO: setting up package sources for latest GNU R installation ...\n"
+    sudo apt-key add ./cran40/jranke.asc && \
+    sudo cp ./cran40/gnu-r.list /etc/apt/sources.list.d/ && \
+    sudo apt update && \
+    echo -e "\nINFO: installing latest GNU R ...\n" && \
+    apt install -y -t bullseye-cran40 r-base-dev
+fi
 
 
 ## setup auto-cpufreq
@@ -112,4 +130,4 @@ fi
 # sudo systemctl mask plymouth-quit-wait.service
 
 
-## edit /etc/default/grub to change grub timeout, update grub by running 'sudo update-grub' 
+## edit /etc/default/grub to change grub timeout, update grub by running 'sudo update-grub'
